requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Django==3.2.11
django-environ==0.8.1
django-extra-fields==3.0.2
django-lifecycle==0.9.3
django-cors-headers==3.10.1
djangorestframework==3.13.1
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
urlman==2.0.1
zope.interface==5.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-environ==0.8.1
django-geojson==3.2.0
djangorestframework-gis==0.18
requests==2.27.1
django-json-widget==1.1.1
django-map-widgets==0.3.2
googlemaps==4.5.3
django-leaflet==0.28.2
xlrd==2.0.1
xlwt==1.3.0
redis==4.1.0
celery==5.1.2
django-celery-results==2.2.0
django-celery-beat==2.2.1
boto3==1.20.34
botocore==1.23.34
django-admin-rangefilter==0.5.4
django-admin-views==0.8.0
Pillow==8.0.0
django-extensions==3.1.5
uWSGI==2.0.20
django-admin-interface
django-rest-auth==0.9.5
factory-boy==3.2.1
Faker==9.8.2